enhanced quicksearch bar
 * inline searchprovider selection, part of faceted /search plugin plan
 * replace ugly standard submit button
 * provide helper text
 * (questionable) trac:#9643 live search integration

getting the quicksearch icon to line up across different browsers is a
true exercise in futility. L56 in this patch works for firefox, but
IE, Safari, Chrome are all of by something like a single pixel.
it seems it probably makes sense to use a transparently backgrounded
non-framed icon instead, so the single-pixel rendering differences are
mostly hidden to the naked eye.